requirements.txt
flask
flask-socketio
flask-cors
numpy
Pillow
python-engineio
python-socketio